requirements.txt
# packages in environment at /usr/local/Caskroom/miniforge/base/envs/agrifinance:
#
# Name                    Version                   Build  Channel
anyio                     4.8.0              pyhd8ed1ab_0    conda-forge
asgiref                   3.8.1              pyhd8ed1ab_1    conda-forge
astroid                   3.3.8           py312hb401068_0    conda-forge
brotli-python             1.1.0           py312h5861a67_2    conda-forge
bzip2                     1.0.8                hfdf4475_7    conda-forge
ca-certificates           2025.1.31            h8857fd0_0    conda-forge
certifi                   2024.12.14         pyhd8ed1ab_0    conda-forge
cffi                      1.17.1          py312hf857d28_0    conda-forge
charset-normalizer        3.4.1              pyhd8ed1ab_0    conda-forge
colorama                  0.4.6              pyhd8ed1ab_1    conda-forge
dill                      0.3.9              pyhd8ed1ab_1    conda-forge
django                    5.1.6              pyhd8ed1ab_0    conda-forge
django-stubs              5.1.2              pyhd8ed1ab_0    conda-forge
django-stubs-ext          5.1.2              pyhd8ed1ab_0    conda-forge
djangorestframework       3.15.2             pyhd8ed1ab_1    conda-forge
exceptiongroup            1.2.2              pyhd8ed1ab_1    conda-forge
h11                       0.14.0             pyhd8ed1ab_1    conda-forge
h2                        4.2.0              pyhd8ed1ab_0    conda-forge
hpack                     4.1.0              pyhd8ed1ab_0    conda-forge
httpcore                  1.0.7              pyh29332c3_1    conda-forge
httpx                     0.28.1             pyhd8ed1ab_0    conda-forge
hyperframe                6.1.0              pyhd8ed1ab_0    conda-forge
idna                      3.10               pyhd8ed1ab_1    conda-forge
iniconfig                 2.0.0              pyhd8ed1ab_1    conda-forge
isort                     6.0.0              pyhd8ed1ab_0    conda-forge
joblib                    1.4.2              pyhd8ed1ab_1    conda-forge
libblas                   3.9.0           28_h7f60823_openblas    conda-forge
libcblas                  3.9.0           28_hff6cab4_openblas    conda-forge
libcxx                    19.1.7               hf95d169_0    conda-forge
libexpat                  2.6.4                h240833e_0    conda-forge
libffi                    3.4.2                h0d85af4_5    conda-forge
libgfortran               5.0.0           13_2_0_h97931a8_3    conda-forge
libgfortran5              13.2.0               h2873a65_3    conda-forge
liblapack                 3.9.0           28_h236ab99_openblas    conda-forge
liblzma                   5.6.4                hd471939_0    conda-forge
libopenblas               0.3.28          openmp_hbf64a52_1    conda-forge
libsqlite                 3.48.0               hdb6dae5_1    conda-forge
libzlib                   1.3.1                hd23fc13_2    conda-forge
llvm-openmp               19.1.7               ha54dae1_0    conda-forge
mccabe                    0.7.0              pyhd8ed1ab_1    conda-forge
ncurses                   6.5                  h0622a9a_3    conda-forge
numpy                     2.2.2           py312h6693b03_0    conda-forge
openssl                   3.4.1                hc426f3f_0    conda-forge
packaging                 24.2               pyhd8ed1ab_2    conda-forge
pandas                    2.2.3           py312h98e817e_1    conda-forge
pip                       25.0               pyh8b19718_0    conda-forge
platformdirs              4.3.6              pyhd8ed1ab_1    conda-forge
pluggy                    1.5.0              pyhd8ed1ab_1    conda-forge
pycparser                 2.22               pyh29332c3_1    conda-forge
pylint                    3.3.4              pyh29332c3_0    conda-forge
pysocks                   1.7.1              pyha55dd90_7    conda-forge
pytest                    8.3.4              pyhd8ed1ab_1    conda-forge
pytest-asyncio            0.25.3             pyh29332c3_0    conda-forge
python                    3.12.8          h9ccd52b_1_cpython    conda-forge
python-dateutil           2.9.0.post0        pyhff2d567_1    conda-forge
python-dotenv             1.0.1              pyhd8ed1ab_1    conda-forge
python-tzdata             2025.1             pyhd8ed1ab_0    conda-forge
python_abi                3.12                    5_cp312    conda-forge
pytz                      2024.1             pyhd8ed1ab_0    conda-forge
readline                  8.2                  h9e318b2_1    conda-forge
requests                  2.32.3             pyhd8ed1ab_1    conda-forge
scikit-learn              1.6.1           py312he1a5313_0    conda-forge
scipy                     1.15.1          py312hb4e66ee_0    conda-forge
setuptools                75.8.0             pyhff2d567_0    conda-forge
six                       1.17.0             pyhd8ed1ab_0    conda-forge
sniffio                   1.3.1              pyhd8ed1ab_1    conda-forge
sqlparse                  0.5.3              pyhd8ed1ab_0    conda-forge
threadpoolctl             3.5.0              pyhc1e730c_0    conda-forge
tk                        8.6.13               h1abcd95_1    conda-forge
tomli                     2.2.1              pyhd8ed1ab_1    conda-forge
tomlkit                   0.13.2             pyha770c72_1    conda-forge
types-pyyaml              6.0.12.20241230    pyhd8ed1ab_0    conda-forge
typing_extensions         4.12.2             pyha770c72_1    conda-forge
tzdata                    2025a                h78e105d_0    conda-forge
urllib3                   2.3.0              pyhd8ed1ab_0    conda-forge
wheel                     0.45.1             pyhd8ed1ab_1    conda-forge
zstandard                 0.23.0          py312h7122b0e_1    conda-forge
zstd                      1.5.6                h915ae27_0    conda-forge
